# openapi.json
{
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ScheduleCard": {
        "properties": {
          "medicine_datetime": {
            "format": "date-time",
            "title": "Medicine Datetime",
            "type": "string"
          },
          "medicine_name": {
            "title": "Medicine Name",
            "type": "string"
          }
        },
        "required": [
          "medicine_name",
          "medicine_datetime"
        ],
        "title": "ScheduleCard",
        "type": "object"
      },
      "ScheduleCreate": {
        "properties": {
          "intake_finish": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "\u041a\u043e\u043d\u0435\u0446 \u043b\u0435\u0447\u0435\u043d\u0438\u044f, null - \u0435\u0441\u043b\u0438 \u043d\u0435\u0442 \u043e\u0433\u0440\u0430\u043d\u0438\u0447\u0435\u043d\u0438\u044f \u0432 \u0434\u043b\u0438\u0442\u0435\u043b\u044c\u043d\u043e\u0441\u0442\u0438 \u043b\u0435\u0447\u0435\u043d\u0438\u044f",
            "title": "Intake Finish"
          },
          "intake_period": {
            "description": "\u041f\u0435\u0440\u0438\u043e\u0434 \u043f\u0440\u0438\u0451\u043c\u043e\u0432 \u0437\u0430\u043f\u0438\u0441\u044b\u0432\u0430\u0435\u0442\u0441\u044f \u0432 [cron \u0441\u0438\u043d\u0442\u0430\u043a\u0441\u0438\u0441\u0435](https://en.wikipedia.org/wiki/Cron#CRON_expression), \u043f\u0440\u0438\u043c\u0435\u0440 0 12 * * * - \u043a\u0430\u0436\u0434\u044b\u0439 \u0434\u0435\u043d\u044c \u0432 \u0440\u043e\u0432\u043d\u043e 12 \u0447\u0430\u0441\u043e\u0432 \u0434\u043d\u044f. **W \u0438 # \u0441\u0438\u043c\u0432\u043e\u043b\u044b \u043d\u0435 \u043f\u043e\u0434\u0434\u0435\u0440\u0436\u0438\u0432\u0430\u044e\u0442\u0441\u044f**",
            "examples": [
              "0 12 * * *"
            ],
            "title": "Intake Period",
            "type": "string"
          },
          "intake_start": {
            "default": "2025-05-08T18:53:08.620660Z",
            "format": "date-time",
            "title": "Intake Start",
            "type": "string"
          },
          "medicine_name": {
            "title": "Medicine Name",
            "type": "string"
          },
          "user_id": {
            "title": "User Id",
            "type": "integer"
          }
        },
        "required": [
          "medicine_name",
          "intake_period",
          "user_id",
          "intake_finish"
        ],
        "title": "ScheduleCreate",
        "type": "object"
      },
      "ScheduleRead": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          }
        },
        "required": [
          "id"
        ],
        "title": "ScheduleRead",
        "type": "object"
      },
      "TakingsRead": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "medicine_datetime": {
            "format": "date-time",
            "title": "Medicine Datetime",
            "type": "string"
          },
          "medicine_name": {
            "title": "Medicine Name",
            "type": "string"
          }
        },
        "required": [
          "medicine_name",
          "medicine_datetime",
          "id"
        ],
        "title": "TakingsRead",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/next_takings": {
      "get": {
        "description": "\u0412\u043e\u0437\u0432\u0440\u0430\u0449\u0430\u0435\u0442 \u0434\u0430\u043d\u043d\u044b\u0435 \u043e \u0442\u0430\u0431\u043b\u0435\u0442\u043a\u0430\u0445, \u043a\u043e\u0442\u043e\u0440\u044b\u0435 \u043d\u0435\u043e\u0431\u0445\u043e\u0434\u0438\u043c\u043e \u043f\u0440\u0438\u043d\u044f\u0442\u044c\n\u0432 \u0431\u043b\u0438\u0436\u0430\u0439\u0448\u0438\u0435 \u043f\u0435\u0440\u0438\u043e\u0434 (\u043d\u0430\u043f\u0440\u0438\u043c\u0435\u0440, \u0432 \u0431\u043b\u0438\u0436\u0430\u0439\u0448\u0438\u0439 \u0447\u0430\u0441). \u041f\u0435\u0440\u0438\u043e\u0434\n\u0432\u0440\u0435\u043c\u0435\u043d\u0438 \u0437\u0430\u0434\u0430\u0435\u0442\u0441\u044f \u0447\u0435\u0440\u0435\u0437 \u043f\u0430\u0440\u0430\u043c\u0435\u0442\u0440\u044b \u043a\u043e\u043d\u0444\u0438\u0433\u0443\u0440\u0430\u0446\u0438\u0438 \u0441\u0435\u0440\u0432\u0438\u0441\u0430",
        "operationId": "next_takings_next_takings_get",
        "parameters": [
          {
            "in": "query",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-Trace-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Trace-Id"
            }
          },
          {
            "in": "header",
            "name": "X-Request-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Request-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TakingsRead"
                  },
                  "title": "Response Next Takings Next Takings Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Next Takings",
        "tags": [
          "Schedule"
        ]
      }
    },
    "/schedule": {
      "get": {
        "description": "\u0412\u043e\u0437\u0432\u0440\u0430\u0449\u0430\u0435\u0442 \u0434\u0430\u043d\u043d\u044b\u0435 \u043e \u0432\u044b\u0431\u0440\u0430\u043d\u043d\u043e\u043c \u0440\u0430\u0441\u043f\u0438\u0441\u0430\u043d\u0438\u0438 \u0441 \u0440\u0430\u0441\u0441\u0447\u0438\u0442\u0430\u043d\u043d\u044b\u043c\n\u0433\u0440\u0430\u0444\u0438\u043a\u043e\u043c \u043f\u0440\u0438\u0451\u043c\u043e\u0432 \u043d\u0430 \u0434\u0435\u043d\u044c",
        "operationId": "schedule_schedule_get",
        "parameters": [
          {
            "in": "query",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "schedule_id",
            "required": true,
            "schema": {
              "title": "Schedule Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-Trace-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Trace-Id"
            }
          },
          {
            "in": "header",
            "name": "X-Request-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Request-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ScheduleCard"
                  },
                  "title": "Response Schedule Schedule Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Bad Request",
            "detail": "\u0420\u0430\u0441\u043f\u0438\u0441\u0430\u043d\u0438\u0435 \u0438\u0441\u0442\u0435\u043a\u043b\u043e!"
          },
          "404": {
            "description": "Not Found",
            "detail": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Schedule",
        "tags": [
          "Schedule"
        ]
      },
      "post": {
        "operationId": "create_schedule_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Trace-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Trace-Id"
            }
          },
          {
            "in": "header",
            "name": "X-Request-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Request-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create",
        "tags": [
          "Schedule"
        ]
      }
    },
    "/schedules": {
      "get": {
        "description": "\u0412\u043e\u0437\u0432\u0440\u0430\u0449\u0430\u0435\u0442 \u0434\u0430\u043d\u043d\u044b\u0435 \u043e \u0432\u044b\u0431\u0440\u0430\u043d\u043d\u043e\u043c \u0440\u0430\u0441\u043f\u0438\u0441\u0430\u043d\u0438\u0438 \u0441 \u0440\u0430\u0441\u0441\u0447\u0438\u0442\u0430\u043d\u043d\u044b\u043c\n\u0433\u0440\u0430\u0444\u0438\u043a\u043e\u043c \u043f\u0440\u0438\u0451\u043c\u043e\u0432 \u043d\u0430 \u0434\u0435\u043d\u044c",
        "operationId": "schedules_schedules_get",
        "parameters": [
          {
            "in": "query",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-Trace-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Trace-Id"
            }
          },
          {
            "in": "header",
            "name": "X-Request-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Request-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ScheduleRead"
                  },
                  "title": "Response Schedules Schedules Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Schedules",
        "tags": [
          "Schedule"
        ]
      }
    }
  }
}